Big restructure

+++ Dance/Graphics/Shader.h
@@ -24,7 +24,7 @@ public:
         ComPtr<ID3D11Device> device,
         const WCHAR* path,
         const D3D11_INPUT_ELEMENT_DESC* layout,
-        size_t count
+        UINT count
     )
     {
         device->GetImmediateContext(this->deviceContext.ReleaseAndGetAddressOf());
